Resume-App/Backend-Testing/Interview/resume-analyser-interview/requirements.txt
openai-whisper 
PyAudio 
python-multipart
uvicorn 
fastapi
dotenv
google-genai